requirements.txt
flask
transformers
torch
psycopg2-binary